Restructuring for some berivity.
