+++ src/query.cpp
@@ -1,0 +1,1 @@
+#include "query.hpp"
